Fixed tests [demos]
